made logScreen methods private -- Assembla ticket #67

+++ src/com/mparticle/MessageManager.java
@@ -324,7 +324,7 @@ import com.mparticle.MParticleAPI.EventType;
         mUploadHandler.setConnectionScheme(sslEnabled ? "https" : "http");
     }
 
-    public void setUploadInterval(int uploadInterval) {
+    public void setUploadInterval(long uploadInterval) {
         mUploadHandler.setUploadInterval(uploadInterval);
     }
